+++ packages/infra/_src/services/Store/utils.ts
@@ -79,12 +79,16 @@ export function codeFilter<E extends { id: string }, NE extends E>(filter: Filte
                 return gtCaseInsensitive(k, p.value)
               case "gte":
                 return gteCaseInsensitive(k, p.value)
+              case "includes":
+                return (k as Array<string>).includes(p.value)
+              case "not-includes":
+                return !(k as Array<string>).includes(p.value)
               case "contains":
-                return (k as string).toLowerCase().includes((p.value as string).toLowerCase())
+                return (k as string).toLowerCase().includes((p.value).toLowerCase())
               case "ends-with":
-                return (k as string).toLowerCase().endsWith((p.value as string).toLowerCase())
+                return (k as string).toLowerCase().endsWith((p.value).toLowerCase())
               case "starts-with":
-                return (k as string).toLowerCase().startsWith((p.value as string).toLowerCase())
+                return (k as string).toLowerCase().startsWith((p.value).toLowerCase())
               case "not-eq":
                 return !compareCaseInsensitive(k, p.value)
               case "eq":
@@ -111,12 +115,16 @@ export function codeFilter<E extends { id: string }, NE extends E>(filter: Filte
                 return gtCaseInsensitive(k, p.value)
               case "gte":
                 return gteCaseInsensitive(k, p.value)
+              case "includes":
+                return (k as Array<string>).includes(p.value)
+              case "not-includes":
+                return !(k as Array<string>).includes(p.value)
               case "contains":
-                return (k as string).toLowerCase().includes((p.value as string).toLowerCase())
+                return (k as string).toLowerCase().includes((p.value).toLowerCase())
               case "ends-with":
-                return (k as string).toLowerCase().endsWith((p.value as string).toLowerCase())
+                return (k as string).toLowerCase().endsWith((p.value).toLowerCase())
               case "starts-with":
-                return (k as string).toLowerCase().startsWith((p.value as string).toLowerCase())
+                return (k as string).toLowerCase().startsWith((p.value).toLowerCase())
               case "not-eq":
                 return p.key.includes(".-1.")
                   ? (get(x, p.key.split(".-1.")[0]) as any[])