+++ packages/infra-adapters/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @effect-app/infra-adapters
 
+## 0.49.0
+
+### Minor Changes
+
+- housekeeping: update packages
+
+### Patch Changes
+
+- Updated dependencies
+  - @effect-app/prelude@0.58.0
+  - @effect-app/schema@0.50.0
+  - @effect-app/core@0.56.0
+
 ## 0.48.2
 
 ### Patch Changes