+++ packages/react/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @effect-app/react
 
+## 0.41.0
+
+### Minor Changes
+
+- housekeeping: update packages
+
+### Patch Changes
+
+- Updated dependencies
+  - @effect-app/prelude@0.58.0
+  - @effect-app/schema@0.50.0
+  - @effect-app/core@0.56.0
+
 ## 0.40.2
 
 ### Patch Changes